+++ BackEnd/loginmicroservices/src/main/java/com/rmit/sept/bk_loginservices/web/RequestController.java
@@ -29,7 +29,6 @@ public class RequestController {
     @Autowired
     private MapValidationErrorService mapValidationErrorService;
 
-
     @GetMapping(path = "/all")
     public Iterable<Request> getAllRequests() {
         return requestService.findAllRequests();
@@ -37,12 +36,18 @@ public class RequestController {
 
     @DeleteMapping(path = "/{requestId}")
     public ResponseEntity<?> deleteRequest(@PathVariable Long requestId) {
-        requestService.deleteRequestById(requestId);
-        return new ResponseEntity<String>("Request with ID " + requestId + " was deleted", HttpStatus.OK);
+        Request request = requestService.findById(requestId);
+
+        if (request != null) {
+            requestService.deleteRequestById(requestId);
+            return new ResponseEntity<String>("Request with ID " + requestId + " was deleted", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<String>("Request with ID " + requestId + " was not found", HttpStatus.ACCEPTED);
+        }
     }
 
     @PostMapping("/new")
-    public ResponseEntity<?> addNewBook(@Valid @RequestBody Request request, BindingResult result) {
+    public ResponseEntity<?> addNewRequest(@Valid @RequestBody Request request, BindingResult result) {
         ResponseEntity<?> errorMap = mapValidationErrorService.MapValidationService(result);
         if (errorMap != null)
             return errorMap;
@@ -51,10 +56,9 @@ public class RequestController {
         if (newRequest != null) {
             return new ResponseEntity<Request>(newRequest, HttpStatus.CREATED);
         } else {
-            return new ResponseEntity<String>("Unable to save details for book, a copy of the book already exists.",
+            return new ResponseEntity<String>("Unable to add the new request, a copy of the request already exists.",
                     HttpStatus.ACCEPTED);
         }
     }
-
 
 }
